+++ src/assets/radio_check.png
