+++ src/ui/screens/details/components/__test__/TransationId.test.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import "@testing-library/jest-native";
 import { render, fireEvent } from "@testing-library/react-native";
-import { transactionFixture } from "../../../../../../__test__/fixtures/transactions";
-import DetailScreen, { DetailScreenProps } from "../../DetailsScreen";
 import * as Clipboard from "expo-clipboard";
-import TransactionId from "../TransactionId";
+import TransactionId from "@/ui/screens/details/components/TransactionId";
 
 describe("<TransactionId />", () => {
     test("Transaction ID copied on press", () => {
