+++ src/ui/screens/index/components/__test__/FilterModal.test.tsx
@@ -1,10 +1,7 @@
 import React from "react";
-import { TransactionCard } from "../TransactionCard";
 import "@testing-library/jest-native";
-import { transactionFixture } from "../../../../../../__test__/fixtures/transactions";
 import { render, fireEvent } from "@testing-library/react-native";
-import { NavigationContainer } from "@react-navigation/native";
-import FilterModal, { SortBy } from "../FilterModal";
+import FilterModal, { SortBy } from "@/ui/screens/index/components/FilterModal";
 
 describe("<FilterModal />", () => {
     test("Should select the pressed value", () => {
