+++ src/ui/screens/index/components/__test__/TransactionCard.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { TransactionCard } from "../TransactionCard";
+import { TransactionCard } from "@/ui/screens/index/components/TransactionCard";
 import "@testing-library/jest-native";
-import { transactionFixtures } from "../../../../../../__test__/fixtures/transactions";
+import { transactionFixtures } from "@/__test__/fixtures/transactions";
 import { render } from "@testing-library/react-native";
 import { NavigationContainer } from "@react-navigation/native";
